app/src/main/java/com/example/unittestcase/model/Note.kt
package com.example.unittestcase.model

import android.os.Parcelable
import androidx.annotation.NonNull
import androidx.room.ColumnInfo
import androidx.room.Entity
import androidx.room.PrimaryKey
import kotlinx.android.parcel.Parcelize

@Parcelize
@Entity(tableName = "notes")
data class Note(
    @PrimaryKey(autoGenerate = true)
    val id: Int? = null,
    @ColumnInfo(name = "title")
    val title: String,
    @ColumnInfo(name = "content")
    val content: String,
    @ColumnInfo(name = "timestamp")
    val timeStamp: String
) : Parcelable {
    override fun equals(other: Any?): Boolean {
        val note = other as Note
        return note.id == id && note.title == title && note.content == content
    }
}